Missile followup
* If homing delayed or disabled, launch parameters chosen to emulate a ballistic projectile (bullet)
* If homing enabled with finite range limit (finite fuel) and not exploding when out of fuel, and target not in homing range, the missile will try to position itself so as to freefall onto the target after the fuel runs out
* Variable names more readable
* More smaller functions instead of few big ones
* Retested SAMs, submarines, adv. guard towers
* Made lookahead more precise and less burdensome (polling more densely, but much less often)

+++ OpenRA.Game/WAngle.cs
@@ -73,10 +73,10 @@ namespace OpenRA
 				return new WAngle(Math.Sign(y) * 256);
 
 			var ay = Math.Abs(y);
-			var ax = Math.Abs(x);
+			var ax = (long)Math.Abs(x);
 
 			// Find the closest angle that satisfies y = x*tan(theta)
-			var bestVal = int.MaxValue;
+			var bestVal = (long)int.MaxValue;
 			var bestAngle = 0;
 			for (var i = 0; i < 256; i += stride)
 			{
